lightning/livekit_example/requirements.txt
livekit-agents>=0.10.1
livekit-plugins-openai>=0.10.3
livekit-plugins-deepgram>=0.6.8
livekit-plugins-silero>=0.7.1
python-dotenv~=1.0
aiofile~=3.8.8